refactor: Port-Adapter + DTO 下沉 + 架构违规修复
Port-Adapter:
- AiGateway interface: infrastructure → domain (proper hexagonal port)
- AiChatRequest/AiChatResponse: infrastructure.ai.dto → domain.ai
- AiGatewayImpl updated to import from domain

DTO 下沉 (fix application → interfaces violations):
- WatermarkConfigRequest → application.shared.dto
- AlbumCreateRequest, AlbumResponse → application.shared.dto
- AiStatsResponse → application.shared.dto
- CreateExcalidrawSceneRequest, ExcalidrawSceneResponse → application.shared.dto
- AdminUserSummary → application.shared.dto
- CommentCreateRequest, CommentResponse → application.shared.dto
- PictureVersionResponse → application.shared.dto

15 files updated with new import paths.
26 tests passing.

Closes #24
