+++ src/main/java/com/cn/cloudpictureplatform/application/shared/dto/CreateExcalidrawSceneRequest.java
@@ -1,4 +1,4 @@
-package com.cn.cloudpictureplatform.interfaces.excalidraw.dto;
+package com.cn.cloudpictureplatform.application.shared.dto;
 
 import lombok.Data;
