+++ src/main/java/com/cn/cloudpictureplatform/domain/ai/AiChatRequest.java
@@ -1,4 +1,4 @@
-package com.cn.cloudpictureplatform.infrastructure.ai.dto;
+package com.cn.cloudpictureplatform.domain.ai;
 
 import java.util.List;
 import java.util.UUID;
