+++ src/main/java/com/cn/cloudpictureplatform/infrastructure/ai/gateway/AiGatewayImpl.java
@@ -10,8 +10,9 @@ import java.util.UUID;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
-import com.cn.cloudpictureplatform.infrastructure.ai.dto.AiChatRequest;
-import com.cn.cloudpictureplatform.infrastructure.ai.dto.AiChatResponse;
+import com.cn.cloudpictureplatform.domain.ai.AiChatRequest;
+import com.cn.cloudpictureplatform.domain.ai.AiChatResponse;
+import com.cn.cloudpictureplatform.domain.ai.AiGateway;
 import com.cn.cloudpictureplatform.infrastructure.persistence.AiCallAuditRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
