+++ src/main/java/com/cn/cloudpictureplatform/interfaces/excalidraw/ExcalidrawSceneController.java
@@ -3,8 +3,8 @@ package com.cn.cloudpictureplatform.interfaces.excalidraw;
 import com.cn.cloudpictureplatform.application.excalidraw.ExcalidrawSceneService;
 import com.cn.cloudpictureplatform.common.web.ApiResponse;
 import com.cn.cloudpictureplatform.infrastructure.security.AppUserPrincipal;
-import com.cn.cloudpictureplatform.interfaces.excalidraw.dto.CreateExcalidrawSceneRequest;
-import com.cn.cloudpictureplatform.interfaces.excalidraw.dto.ExcalidrawSceneResponse;
+import com.cn.cloudpictureplatform.application.shared.dto.CreateExcalidrawSceneRequest;
+import com.cn.cloudpictureplatform.application.shared.dto.ExcalidrawSceneResponse;
 import com.cn.cloudpictureplatform.interfaces.excalidraw.dto.UpdateSnapshotRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
