+++ src/main/java/com/cn/cloudpictureplatform/interfaces/watermark/WatermarkController.java
@@ -17,7 +17,7 @@ import com.cn.cloudpictureplatform.domain.watermark.ExportPreset;
 import com.cn.cloudpictureplatform.domain.watermark.ExportTask;
 import com.cn.cloudpictureplatform.domain.watermark.WatermarkConfig;
 import com.cn.cloudpictureplatform.infrastructure.security.AppUserPrincipal;
-import com.cn.cloudpictureplatform.interfaces.watermark.dto.WatermarkConfigRequest;
+import com.cn.cloudpictureplatform.application.shared.dto.WatermarkConfigRequest;
 
 @RestController
 @RequestMapping("/api/v1")
